% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_shannon.R
\name{my_shannon}
\alias{my_shannon}
\title{My Shannon diversity function}
\usage{
my_shannon(x)
}
\arguments{
\item{x}{A vector of numeric values representing the species abundance in
a community}
}
\value{
Shannon diversity index
import vegan diversity
}
\description{
My Shannon diversity function
}
\examples{
comm <- c(1, 4, 7, 8)
my_shannon (comm)
}
